Normal connection trace + some post processing.
All the events of a normal TCP connection are logged both on the client
and server sides.
Some post processing is applied on each event to compute seqno and acks
relative to the initial ones. It allows to compact the logs and provides
a format similar to the one of tcpdump.

The code is quite unclean but is functional. Some cleaning will be
applied later.