Add rcv win + bpftool install
Currently the rcv win for send event is 0.
An alternative for the current send kprobes is bpf_skops_write_hdr_opt (https://elixir.bootlin.com/linux/latest/source/net/ipv4/tcp_output.c#L534).
This is a sockops BPF program hook which is not usable with BCC but it should be usable with bpftool and C user-space programs.